feat(product): adiciona validação ao tentar deletar produto informando id não existente

+++ src/main/java/com/br/productmanager/service/impl/ProductServiceImpl.java
@@ -75,6 +75,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void delete(UUID id) {
+        if (!repository.existsById(id)) {
+            throw new ProductNotFoundException("id not found");
+        }
+
         repository.deleteById(id);
     }
 }
